Add files via upload
There are currently three options for tracking the sperms. 
K-means clustering: This will cluster them based on euclidean distance, however any sperm collisions will end up with merged tracks and you have to manually figure out the number of clusters.
DBSCAN: This will cluster them based on being less than a certain distance from each other. However this will often guess the incorrect number of sperm tracks and the hyper parameters are difficult to tune. 
Mike's: This is a physics-based brute force approach. This will cluster the sperm tracks based on the closest sperm in the previous frame. Perfectly tracks sperms, however sperms that leave the frame then rejoin the frame are classified as 'new' sperms. Furthermore, the accuracy can be increased by using more physics that 'closest sperm'. Also assumes that it is impossible for a sperm to leave a frame and new sperm enters the frame.